// src/types/enums/socket-events.ts
export enum SocketEvents {
  CONNECTED_EVENT = "connected",
  DISCONNECTED_EVENT = "disconnected",
  SOCKET_ERROR_EVENT = "socketerror",
  ADMIN_NOTIFICATION_COUNT = "admin-notification-count",
  JOIN_ADMIN_ROOM = "join-admin-room",
  NEW_ADMIN_REQUEST = "new-admin-request",
  JOIN_USER_ROOM = "join-user-room",
  UPDATE_REQUEST_STATUS = "update-request-status",
  REQUEST_STATUS_UPADATE = "request-status-update",
  REQUEST_UPADATED = "request-updated",
  ADMIN_MESSAGE_BROADCAST = "admin-message-broadcast",
  TRANSFER_TRANSACTION = "transfer-transaction",
  DEPOSIT_TRANSACTION = "deposit-transaction",
  DEBIT_TRANSACTION = "debit-transaction",
  WIDTHDRAW_TRANSACTION = "withdraw-transaction",
}